Further reduced the verification method clutter.
