+++ src/ParallelHelper.Test/Analyzer/AnalyzerTestBase.cs
@@ -1,10 +1,5 @@
-﻿using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
+﻿using Microsoft.CodeAnalysis.Diagnostics;
 using System.Collections.Immutable;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace ParallelHelper.Test.Analyzer {
   /// <summary>
@@ -13,15 +8,23 @@ namespace ParallelHelper.Test.Analyzer {
   /// <typeparam name="TAnalyzer">The type of the analyzer under test.</typeparam>
   public class AnalyzerTestBase<TAnalyzer> where TAnalyzer : DiagnosticAnalyzer, new() {
     /// <summary>
+    /// Creates a new compilation builder instance with the tested analyzer.
+    /// </summary>
+    /// <returns>A compilation builder that already contains the tested analyzer.</returns>
+    public TestCompilationBuilder CreateAnalyzerCompilationBuilder() {
+      return TestCompilationBuilder.Create()
+        .AddAnalyzers(new TAnalyzer());
+    }
+
+    /// <summary>
     /// Verifies that the given diagnostics are reported when analyzing the given source.
     /// </summary>
     /// <param name="source">The source to analyze.</param>
     /// <param name="expectedDiagnostics">The expected diagnostics.</param>
     public void VerifyDiagnostic(string source, params DiagnosticResultLocation[] expectedDiagnostics) {
-      var compilationBuilder = TestCompilationBuilder.Create()
+      CreateAnalyzerCompilationBuilder()
         .AddSourceTexts(source)
-        .AddAnalyzers(new TAnalyzer());
-      VerifyDiagnostic(compilationBuilder, expectedDiagnostics);
+        .VerifyDiagnostic(expectedDiagnostics);
     }
 
     /// <summary>
@@ -35,48 +38,10 @@ namespace ParallelHelper.Test.Analyzer {
       ImmutableDictionary<string, string> analyzerOptions,
       params DiagnosticResultLocation[] expectedDiagnostics
     ) {
-      var compilationBuilder = TestCompilationBuilder.Create()
+      CreateAnalyzerCompilationBuilder()
         .AddSourceTexts(source)
-        .AddAnalyzers(new TAnalyzer())
-        .WithAnalyzerOptions(analyzerOptions);
-      VerifyDiagnostic(compilationBuilder, expectedDiagnostics);
-    }
-
-    private void VerifyDiagnostic(TestCompilationBuilder compilationBuilder, DiagnosticResultLocation[] expectedDiagnostics) {
-      var diagnosticResults = Analyze(compilationBuilder);
-      Assert.AreEqual(expectedDiagnostics.Length, diagnosticResults.Length, "Invalid diagnostics count");
-
-      for(var i = 0; i < expectedDiagnostics.Length; ++i) {
-        var result = diagnosticResults[i];
-        var expected = expectedDiagnostics[i];
-        var span = result.Location.GetLineSpan();
-        var actual = new DiagnosticResultLocation(span.Path, span.StartLinePosition.Line, span.StartLinePosition.Character + 1);
-        Assert.AreEqual(expected, actual, "Invalid diagnostic");
-      }
-    }
-
-    private ImmutableArray<Diagnostic> Analyze(TestCompilationBuilder compilationBuilder) {
-      var diagnostics = Task.Run(async () => await GetDiagnosticsAsync(compilationBuilder)).Result;
-      foreach(var compilationDiagnostic in diagnostics.Compilation) {
-        Console.WriteLine(compilationDiagnostic);
-      }
-      return diagnostics.Analyzer;
-    }
-
-    private async Task<(ImmutableArray<Diagnostic> Analyzer, ImmutableArray<Diagnostic> Compilation)> GetDiagnosticsAsync(
-      TestCompilationBuilder compilationBuilder
-    ) {
-      var analyzerDiagnostics = await compilationBuilder
-        .BuildWithAnalyzers()
-        .GetAnalyzerDiagnosticsAsync();
-      // The result is ordered for better reproducability.
-      return (
-        analyzerDiagnostics.OrderBy(diagnostic => diagnostic.Location.SourceSpan).ToImmutableArray(),
-        compilationBuilder.Build().GetDiagnostics()
-          .Where(diagnostic => diagnostic.Severity >= DiagnosticSeverity.Error)
-          .OrderBy(diagnostic => diagnostic.Location.SourceSpan)
-          .ToImmutableArray()
-      );
+        .WithAnalyzerOptions(analyzerOptions)
+        .VerifyDiagnostic(expectedDiagnostics);
     }
   }
 }
